Fix messaging service injection
