+++ pom.xml
@@ -142,6 +142,7 @@
 
     <build>
         <plugins>
+
             <plugin>
                 <groupId>com.atlassian.maven.plugins</groupId>
                 <artifactId>maven-stash-plugin</artifactId>
