+++ src/main/scala/com/dataember/stash/plugin/hook/MyRepositoryHook.scala
@@ -1,18 +1,23 @@
 package com.dataember.stash.plugin.hook
 
 import java.util
+import java.util.Properties
 
 import com.atlassian.bitbucket.hook.HookResponse
 import com.atlassian.bitbucket.hook.repository.{PreReceiveRepositoryHook, RepositoryHookContext, AsyncPostReceiveRepositoryHook}
 import com.atlassian.bitbucket.repository.{RefChange, Repository}
 import com.atlassian.bitbucket.setting.{Settings, SettingsValidationErrors, RepositorySettingsValidator}
+import com.dataember.stash.plugin.msg.{NatsMessagingService, MessagingService}
+import org.nats.Conn
+import org.springframework.beans.factory.annotation.Autowired
 
 import scalaj.http.Http
 
 /**
   * Created by chaospie on 28/11/15.
   */
-class MyRepositoryHook extends  AsyncPostReceiveRepositoryHook with RepositorySettingsValidator {
+class MyRepositoryHook(val messagingService: MessagingService)
+  extends  AsyncPostReceiveRepositoryHook with RepositorySettingsValidator {
 
 
 
@@ -27,9 +32,7 @@ class MyRepositoryHook extends  AsyncPostReceiveRepositoryHook with RepositorySe
   override def postReceive(repositoryHookContext: RepositoryHookContext,
                            collection: util.Collection[RefChange]): Unit = {
 
-    Option[String](repositoryHookContext.getSettings().getString("url"))
-      .filterNot(_.isEmpty) map(s => Http("http://foo.com/add")
-        .postForm(Seq("name" -> "jon", "age" -> "29")).asString)
+    messagingService.publish("Post receive!!!")
   }
 
 }
