+++ src/main/scala/com/dataember/stash/plugin/msg/MessagingService.java
@@ -1,0 +1,8 @@
+package com.dataember.stash.plugin.msg;
+
+/**
+  * Created on 28/11/15.
+  */
+public interface MessagingService {
+  void publish(String msg);
+}
